pyproject.toml
[project]
name = "selectstuff"
version = "0.1.0"
description = "Collect stuff from your browser"
authors = [
    {name = "Lee Miller",email = "dleemiller@protonmail.com"},
    {name =  "Bradley Smith",email = "bssmith925@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
package-mode = false
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "dspy-ai (>=2.5.43,<3.0.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "wordllama (>=0.3.8.post20,<0.4.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "cloudpickle (>=3.1.1,<4.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "opentelemetry-api (>=1.29.0,<2.0.0)",
    "opentelemetry-sdk (>=1.29.0,<2.0.0)",
    "opentelemetry-instrumentation-requests (>=0.50b0,<0.51)",
    "opentelemetry-exporter-otlp-proto-grpc (>=1.29.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.50b0,<0.51)",
    "opentelemetry-instrumentation-sqlalchemy (>=0.50b0,<0.51)",
    "structlog (>=25.1.0,<26.0.0)"
]

[tool.poetry]
packages = [
    {include = "backend"},
    {include = "applications"},
]

[tool.poetry.group.backend.dependencies]
uvicorn = "^0.34.0"
fastapi = "^0.115.8"

[tool.poetry.group.ui.dependencies]
streamlit = "^1.41.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0,<8.0.0"
pytest-cov = ">=4.1.0,<5.0.0"
pytest-mock = ">=3.11.0,<4.0.0"
tox = ">=4.10.0,<5.0.0"
coverage = ">=7.3.2,<8.0.0"
mypy = ">=1.6.0,<2.0.0"
flake8 = ">=6.1.0,<7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source = [
    "backend",
    "applications",
    "review",
    "databases",
]
branch = true

[tool.coverage.report]
fail_under = 90
skip_covered = true
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.pytest.ini_options]
minversion = "7.4.0"
addopts = "--cov=app --cov=search --cov=review --cov=database --cov=main --cov-report=term-missing --cov-fail-under=90"
testpaths = [
    "tests",
]

[tool.tox]
envlist = ["py{3.12}", "mypy", "flake8", "pytest"]

[tool.tox.environment.pytest]
description = "Run tests and collect coverage"
deps = [
    "pytest",
    "pytest-cov",
]
commands = [
    "pytest",
]

[tool.tox.environment.mypy]
description = "Run type checking with mypy"
deps = ["mypy"]
commands = ["mypy ."]

[tool.tox.environment.flake8]
description = "Lint code using flake8"
deps = ["flake8"]
commands = ["flake8 ."]

